[Continues #150792007] Removing access to internal sets.
Added `GossipDetector` access API to manage internal
sets of "alive" and "suspected" servers:

- `GetRandomNeighbor()`;
- `AddAlive()`;
- `ReportSuspected()`; and
- `RemoveSuspected()`.

This API removes the need to access the internal state of
the sets, which are now fully protected behind mutex locks.

A read-only snapshot of the state of the sets can be retrieved
via the `PrepareReport()` method.

`TODO`: remove all access to internal state and complete refactor.