Updated Consisten Hash logic
More consistent API, also fixed a potential issue with using `float` as keys in the partition points map; this now uses a custom `Comparator` (`FloatLessWithTolerance`) which allows for a custom tolerance around the actual float value.

Added a convenience method to create a `View` and clarified the issue around deleting the `Bucket` pointers in the `Map`.